+++ NEWS.md
@@ -7,6 +7,7 @@
 * Attempting to rename columns using `group_by()` now leads to an error (#799)
 * `pmap()` family works with data frame inputs (#803)
 * `filter()` properly handles when comparing to `NA` when `.by` is used (#812)
+* `paged.print` has been removed since it was breaking console printing (#810)
 
 # tidytable 0.11.0